+++ mobile/android/android-components/docs/changelog.md
@@ -20,7 +20,8 @@ permalink: /changelog/
    * ⚠️ **This is a breaking API change**: Timespan and timing distribution
      metrics now have a thread-safe API. See `adding-new-metrics.md` for more
      information.
-     
+   * `Glean.sendPings` has been added for sending custom pings.
+
 # 0.50.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v0.49.0...v0.50.0)